spa: ffi: loop: Ensure invoke function lives long enough
We need to ensure the function lifetime with a Box into/from_raw. The
Pin isn't necessary. Because it's a Box<dyn FnMut>, which we can't do a
from_raw on (being a fat pointer), we add another Box around it which
allows us to have a thin pointer we can send through the FFI boundary.