pipewire: main_loop: Rework structure of inner
We expect this to be a common pattern, where the exposed structure is a
simple wrapper over a ref-counted inner. The use of `running` in the
outer structure exposed a bunch of incorrect behaviour, which is now
fixed. The inner structure is now where everything lives, as expected.

The support being here is a bit janky -- maybe we can factor this out
into some sort of global functionality.

Doing it this way also allows us to drop the superfluous `Clone` on the
inner.