spa: Mark all callback types as Send
We want to be able to send things between application and main loop
threads. We have an escape hatch in that we explicitly mark
refcounted!() objects as Send, and will need to make sure clients don't
end up using them across threads. Not sure this is _quite_ right, but
let's see how this goes.