spa: pod: Improve object parser aesthetics
The type we are expecting is available in pop_object(), but I think I
didn't know how to carry the type into the ObjectParser when I wrote
this. Now, we carry that type into the parser with a phantom type that
we can use while popping the object.

Also, with this we can implement Interator on the object parser quite
easily.